fix: url escape enroll secret in profile url
